# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fraunhofer IESE",
        "version": "1.0.0"
    },
    "paths": {
        "/": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Students",
                "description": "Returns list of projects",
                "operationId": "getProjectsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Students",
                "description": "Returns project data",
                "operationId": "storeProject",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Student"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Student": {
                "title": "Student",
                "description": "Student model",
                "properties": {
                    "studentId": {
                        "title": "studentId",
                        "description": "studentId",
                        "type": "string"
                    },
                    "fullName": {
                        "title": "FullName",
                        "description": "Full name of student",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "description": "Email of student",
                        "type": "string"
                    },
                    "address": {
                        "title": "address",
                        "description": "address",
                        "type": "string"
                    },
                    "age": {
                        "title": "Age",
                        "description": "Student age's user model",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Student"
                }
            }
        }
    },
    "security": [
        []
    ]
}